virstats: Gathering net interface stats with ovs
When vhostuser or ovs interfaces are used, the interface statistics
are not always available in /proc/net/dev.

This change looks at the openvswitch interfaces statistics
tables to provide this information in additional to /proc/net/dev.

Note that in openvswitch world drop/error doesn't always make sense
for some interface type. When these informations are not available we
set them to 0 on the virDomainInterfaceStats.